Fix 2 bugs and optimize code for mergeFraction.
Detail:

bug1:
Added a judgment, that is, when the input does not contain a fraction, it will be returned as it is.
The previous implementation returned a fraction regardless of whether the input contained a fractional monomial.
This will cause the external function preprocess to fall into an infinite loop.
For example, if the input is 'x+y', the function mergeFraction, which previously incorrectly returned '(x+y)/1', now returns '(x+y)'

bug2:
line 376: numeratorTmp = getNumerator(exprTmp_i);
changed to:
line 377: numeratorTmp = std::move(exprTmp_i);