Refine the project code
Implement combineConstant to complete the merging of constants in expressions
Add calls to sortExpr function to reduce redundancy
Add rewriting of the original expressions
Improve the mergeFraction function
Add the function of sympy optimization function.
Implement the sympy optimization function simplifyExprFromStr, where the input and output are strings without going through the file.